+++ modules/infrastructure/cluster/00-versions.tf
@@ -16,10 +16,6 @@ terraform {
       source  = "hashicorp/null"
       version = ">= 3.0"
     }
-    shell = {
-      source  = "scottwinkler/shell"
-      version = ">= 1.7.10"
-    }
     random = {
       source  = "hashicorp/random"
       version = ">= 3.0"