+++ modules/infrastructure/iam/50-cert-manager-iam.tf
@@ -47,7 +47,8 @@ resource "aws_iam_policy" "cert_manager" {
 resource "aws_iam_role" "cert_manager" {
   count = local.persists_through_sleep && var.enable_cert_manager_iam ? 1 : 0
 
-  name = "${var.cluster_name}-rosa-cert-manager"
+  name                 = substr("${var.cluster_name}-rosa-cert-manager", 0, 64)
+  permissions_boundary = var.custom_permissions_boundary_arn
 
   assume_role_policy = jsonencode({
     Version = "2012-10-17"